Extract_Text_from_Image/requirements.txt
streamlit
opencv-python
pytesseract
Pillow
numpy